requirements.txt
pillow
tqdm
ultralytics
transformers
accelerate
opencv-python